+++ rdm-redis-imp/src/main/java/xyz/hashdog/rdm/redis/imp/client/JedisPoolClient.java
@@ -98,19 +98,19 @@ public class JedisPoolClient extends AbstractRedisClient implements RedisClient 
      * @return 0:DB[size]
      */
     @Override
-    public Map<Integer, Long> dbSize() {
+    public Map<Integer, Integer> dbSize() {
         return execute(jedis->{
             String info = jedis.info("Keyspace");
-            Map<Integer,Long> map = new LinkedHashMap<>();
+            Map<Integer,Integer> map = new LinkedHashMap<>();
             for (int i = 0; i < 15; i++) {
-                map.put(i,0L);
+                map.put(i,0);
             }
             String[] line = info.split("\r\n");
             for (String row : line) {
                 if(!row.startsWith("db")){
                     continue;
                 }
-                Tuple2<Integer,Long> tu=Util.keyspaceParseDb(row);
+                Tuple2<Integer,Integer> tu=Util.keyspaceParseDb(row);
                 map.put(tu.t1(),tu.t2());
             }
             return map;
